now we can insert one post into the db
note that originally we were assuming the db was named blogdb because docker-compose.yaml had POSTGRES_DB: blogdb. However, for some reason, this wasn't being considered when docker was creating the db so the real name of the db was postgres. I found this issue by running \c inside the image